Use the Third Argument of Register Setting
It allows you to specify a callback function to clean and validate all
your user input before sending it to the database. this is useful for
escaping things and such.  OR, in the case of this plugin, making sure
the settings adhere to a certain idea: the filter character should only
be one character.